requirements.txt
## python 3.12
psycopg2-binary==2.9.9
networkx==3.4.2
thefuzz[speedup]==0.22.1
dotenv==1.0.1
requests==2.32.3        
